# Conflicts:
#	include/chunk_list.h
#	include/keywords.h
#	include/output.h
#	include/unc_tools.h
#	src/chunk_list.cpp
#	src/output.cpp
#	src/parse_frame.h

+++ src/output.cpp
@@ -305,7 +305,7 @@ static void add_char(UINT32 ch)
          write_char(ch);
          if (ch == '\t')
          {
-            cpd.column = next_tab_column(cpd.column);
+            cpd.column = (UINT16)next_tab_column(cpd.column);
          }
          else
          {
@@ -425,7 +425,7 @@ void output_parsed(FILE *pfile)
    fprintf(pfile, "# Line              Tag           Parent          Columns Br/Lvl/pp     Flag   Nl  Text");
    for (chunk_t *pc = chunk_get_head(); pc != NULL; pc = chunk_get_next(pc))
    {
-      fprintf(pfile, "\n# %3zu> %16.16s[%16.16s][%3zu/%3zu/%3d/%3d][%zu/%zu/%zu][%10" PRIx64 "][%zu-%d]",
+      fprintf(pfile, "\n# %3zu> %16.16s[%16.16s][%3zu/%3zu/%3u/%3u][%zu/%zu/%zu][%10" PRIx64 "][%zu-%d]",
               pc->orig_line, get_token_name(pc->type),
               get_token_name(pc->parent_type),
               pc->column, pc->orig_col, pc->orig_col_end, pc->orig_prev_sp,
